servers/image-processing/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "image-mcp-server"
version = "0.1.0"
description = "MCP server for image processing and transformations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "pillow>=10.0.0",
    "boto3>=1.34.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["image_mcp_server*"]
